Move title fetching code to title.rs
